requirements.txt
# Sistema POS Sabrositas v2.0.0 - Dependencias de Producción CON IA
# =================================================================
# Las Arepas Cuadradas - Enterprise Architecture
# Versiones optimizadas para Python 3.13 y producción enterprise
# Actualizado: 2024-12-17 - Librerías Actualizadas y Compatibles

# Framework Principal - Versiones más recientes
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-CORS==5.0.0
Flask-Limiter==3.8.0

# Base de Datos - SQLite integrado (sin PostgreSQL para desarrollo)
SQLAlchemy==2.0.36
# psycopg2-binary==2.9.9  # Solo para producción con PostgreSQL

# Seguridad - Versiones más seguras
bcrypt==4.2.0
PyJWT==2.9.0

# Utilidades - Actualizadas
python-dotenv==1.0.1
python-dateutil==2.9.0
pytz==2024.2
requests==2.32.3

# Logging - Versión más reciente
structlog==24.4.0

# IA y Machine Learning - Python 3.13 Compatible (Actualizadas)
scikit-learn==1.5.2
numpy==2.1.3
scipy==1.14.1
joblib==1.4.2
threadpoolctl==3.5.0

# Procesamiento de texto - Actualizado
nltk==3.9.1

# Testing - Versiones más recientes
pytest==8.3.3
pytest-cov==5.0.0

# Desarrollo - Herramientas actualizadas
black==24.10.0
flake8==7.1.1
isort==5.13.2

# Servidor de producción
gunicorn==23.0.0
Werkzeug==3.0.4

# Cache y Queue Management
redis==5.1.1
celery==5.4.0

# Serialización y Validación
marshmallow==3.23.1

# CLI y Utilidades
click==8.1.7

# Logging estructurado
python-json-logger==2.0.7
Jinja2==3.1.4

# Métricas y Monitoreo
prometheus-client==0.21.0

# Dependencias adicionales para producción con PostgreSQL
# Descomentar para producción:
# psycopg2-binary==2.9.9